profit.py
INFYTQ question
The road transport corporation wants to know if a certain bus route is running on a profit or loss.
assume the given in formation:
price per litre of fuel=70
mileage of the bus in km/litre of fuel=10
price(Rs) per ticket=80

The bus runs on multiple routes having different distance in kms and number of passengers.
Write a function to calculate and return the profit earned(Rs) in each route. Return 01 in case of loss.