handshakes.cpp
It was Raj's first day at school. His teacher Anu asked the students to meet every other student in the class and to introduce about themselves. The teacher asked them to do handshakes when they meet each other.

If there are n number of students in the class then find the total number of handshakes made by the students.